renamed the contants to constants file everywhere
