+++ client/src/components/edit-job-lead-component/edit-job-lead-form/index.jsx
@@ -32,7 +32,7 @@ import { formateDateObjToStr } from "../../../utils/date";
 import ErrorScreenComponent from "../../shared/error-screen-component";
 import { modifyJobLead } from "../../../utils/api";
 import ConfirmDialog from "../../shared/confirm-dialog-component";
-import { JOB_TYPES } from "../../../utils/contants";
+import { JOB_TYPES } from "../../../utils/constants";
 import FormSubmissionErrorDialog from "../../shared/form-submission-error-dialog";
 
 function EditJobLeadFormComponent({ jobLead, setSnackBarMessage }) {
